Changed Constant to be subclass of Distribution
