+++ pymc3/distributions/distribution.py
@@ -8,7 +8,7 @@ from ..model import Model, get_named_nodes, FreeRV, ObservedRV
 from ..vartypes import string_types
 
 __all__ = ['DensityDist', 'Distribution', 'Continuous', 'Discrete',
-           'NoDistribution', 'TensorType', 'draw_values']
+           'NoDistribution', 'Constant', 'ConstantDist', 'TensorType', 'draw_values']
 
 
 class _Unpickling(object):
@@ -114,6 +114,49 @@ class NoDistribution(Distribution):
 
     def logp(self, x):
         return 0
+        
+
+class Constant(Distribution):
+    """
+    Constant log-likelihood.
+
+    Parameters
+    ----------
+    c : float or int
+        Constant parameter.
+    """
+
+    def __init__(self, c, dtype=None, *args, **kwargs):
+        super(Constant, self).__init__(*args, **kwargs)
+        if dtype is None:
+            dtype = theano.config.floatX
+        self.mean = self.median = self.mode = self.c = c = tt.as_tensor_variable(c)
+
+    def random(self, point=None, size=None, repeat=None):
+        c = draw_values([self.c], point=point)[0]
+        dtype = np.array(c).dtype
+
+        def _random(c, dtype=dtype, size=None):
+            return np.full(size, fill_value=c, dtype=dtype)
+
+        return generate_samples(_random, c=c, dist_shape=self.shape,
+                                size=size).astype(dtype)
+
+    def logp(self, value):
+        c = self.c
+        return bound(0, tt.eq(value, c))
+
+    def _repr_latex_(self, name=None, dist=None):
+        if dist is None:
+            dist = self
+        return r'${} \sim \text{{Constant}}()$'.format(name)
+
+
+def ConstantDist(*args, **kwargs):
+    import warnings
+    warnings.warn("ConstantDist has been deprecated. In future, use Constant instead.",
+                  DeprecationWarning)
+    return Constant(*args, **kwargs)
 
 
 class Discrete(Distribution):
